Clamp tick marks to inside edge of clock border.
Currently tick marks are plotted from (50,-5) to (50,0) and transformed
(rotated) around the rim of the clock. This causes them to extend to the
far (outside) edge of the clock's border. Instead, tick marks should end
where the border begins. As the border is 1 unit thick, tick marks need
to be plotted from (50,-5) to (50,1).